+++ config/index.js
@@ -1,1 +1,4 @@
+/**
+ * Barrel that re-exports the application configuration.
+ */
 export * from "../config/app";